Separate out cgp-sync from cgp-async with identical API
