+++ crates/cgp-sync/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name         = "cgp-async"
+name         = "cgp-sync"
 version      = "0.1.0"
 edition      = "2021"
 license      = "Apache-2.0"
